Debug vectors

+++ Vertex.java
@@ -3,7 +3,7 @@ public class Vertex {
     public double y;
     public double z;
 
-    public Vertex(int x, int y, int z) {
+    public Vertex(double x, double y, double z) {
         this.x = x;
         this.y = y;
         this.z = z;
